+++ src/main/java/uk/q3c/krail/core/persist/common/option/DefaultOptionDao.java
@@ -16,7 +16,7 @@ package uk.q3c.krail.core.persist.common.option;
 import com.google.common.collect.ImmutableList;
 import com.google.inject.Inject;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
-import uk.q3c.krail.core.data.OptionElementConverter;
+import uk.q3c.krail.core.data.DataListConverter;
 import uk.q3c.krail.core.data.OptionListConverter;
 import uk.q3c.krail.core.option.Option;
 import uk.q3c.krail.core.option.OptionException;
@@ -32,8 +32,7 @@ import uk.q3c.krail.core.user.profile.RankOption;
 import javax.annotation.Nonnull;
 import java.util.Optional;
 
-import static com.google.common.base.Preconditions.checkArgument;
-import static com.google.common.base.Preconditions.checkNotNull;
+import static com.google.common.base.Preconditions.*;
 
 /**
  * Data Access Object for {@link DefaultInMemoryOptionStore}
@@ -47,11 +46,11 @@ import static com.google.common.base.Preconditions.checkNotNull;
 public class DefaultOptionDao implements OptionDao {
 
     private InMemoryOptionStore optionStore;
-    private OptionElementConverter optionElementConverter;
+    private DataListConverter optionElementConverter;
     private OptionDaoDelegate delegate;
 
     @Inject
-    public DefaultOptionDao(OptionElementConverter optionElementConverter, OptionSource delegateSource) {
+    public DefaultOptionDao(DataListConverter optionElementConverter, OptionSource delegateSource) {
         this.optionElementConverter = optionElementConverter;
         this.delegate = delegateSource.getActiveDao();
     }